include working Azure AD authentication and initial test of pushing from Blender to EdgeDB

+++ src/server/api/fastapi_azure_ad/app/routers/migration.py
@@ -6,8 +6,9 @@ import pathlib
 from http import HTTPStatus
 
 import edgedb
-from fastapi import APIRouter
+from fastapi import APIRouter, Security
 from pydantic import BaseModel
+from app.dependencies import azure_scheme
 
 router = APIRouter()
 
@@ -17,24 +18,28 @@ class MigrationData(BaseModel):
     success: bool
 
 
-@router.post("/migrate", status_code=HTTPStatus.CREATED)
-async def initial_migration(db_name: str) -> MigrationData:
+@router.post("/migrate", dependencies=[Security(azure_scheme)], status_code=HTTPStatus.CREATED)
+async def initial_migration(db_name: str = None) -> MigrationData:
     top_dir = pathlib.Path(__file__).parent.parent
 
     print("Migrating schema to fresh database")
     # Extend wait time to give docker compose setup enough time to start
     client = edgedb.create_client()
-    try:
-        print(f"Dropping database {db_name}")
-        client.execute(f"DROP database {db_name}")
-    except edgedb.errors.UnknownDatabaseError as e:
-        logging.debug(e)
+    if db_name is not None:
+        try:
+            print(f"Dropping database {db_name}")
+            client.execute(f"DROP database {db_name}")
+        except edgedb.errors.UnknownDatabaseError as e:
+            logging.debug(e)
 
-    client.execute(f"CREATE DATABASE {db_name}")
-    client.close()
+        client.execute(f"CREATE DATABASE {db_name}")
+        client.close()
 
-    client = edgedb.create_client(database=db_name)
+        client = edgedb.create_client(database=db_name)
+
     migrations_dir = pathlib.Path(top_dir / "config/schema/basic/dbschema/migrations")
+    if migrations_dir.exists() is False:
+        client.execute('')
     for migration_file in os.listdir(migrations_dir):
         with open(migrations_dir / str(migration_file), "r") as f:
             migration_body = f.read()
